A Farmer Friendly Portal
Agriculture is the noblest occupation of man. With the ever growing need for more productivity in agriculture, there is an urgent need for technological advancements in this field. Through research and interaction with farmers, major problems and glitches have been identified by our team in agriculture. The scope of interaction has been limited to a particular village but the problems identified have a universal scope. Marketing is the primary problem of farmers followed by other obstacles like transportation, lack of technical knowledge and networking. Such problems could be solved by developing various software and apps, giving the problems a blend of computer technology. An online portal for the same is considered as a solution for the same where the suitable connections would be established among the farmers and other consumer entities, making agriculture cost effective, efficient and productive. The ultimate aim of the project is to remove the uncertainty that still persists in this field making agriculture a trustworthy occupation.